#51: implemented signature verification in Linux update script

+++ linux/install-update.sh
@@ -32,7 +32,7 @@ fi
 
 echo "Downloading version information..."
 UPDATE_INFO_LINK=`cat "$NZBOP_APPDIR/webui/package-info.json" | sed -n 's/^.*update-info-link.*: "\(.*\)".*/\1/p'`
-"$NZBOP_APPBIN" -B webget "$NZBOP_TEMPDIR/NZBGET_UPDATE.txt" "$UPDATE_INFO_LINK" 2>/dev/null
+"$NZBOP_APPBIN" -n -B webget "$NZBOP_TEMPDIR/NZBGET_UPDATE.txt" "$UPDATE_INFO_LINK" 2>/dev/null
 if test "$?" != "0"; then
     echo "[ERROR] Download failed, please try again later"
     exit 1
@@ -40,27 +40,49 @@ fi
 
 if test "$NZBUP_BRANCH" = "TESTING"; then
     VER_FIELD="testing-version"
-    DOWNLOAD_FIELD="testing-download"
+    DNL_FIELD="testing-download"
+    SIG_FIELD="testing-signature"
 elif test "$NZBUP_BRANCH" = "STABLE"; then
     VER_FIELD="stable-version"
-    DOWNLOAD_FIELD="stable-download"
+    DNL_FIELD="stable-download"
+    SIG_FIELD="stable-signature"
 else
     echo "[ERROR] Unsupported branch $NZBUP_BRANCH"
     exit 1
 fi
 
 VER=`cat "$NZBOP_TEMPDIR/NZBGET_UPDATE.txt" | sed -n "s/^.*$VER_FIELD.*: \"\(.*\)\".*/\1/p"`
-DOWNLOAD_LINK=`cat "$NZBOP_TEMPDIR/NZBGET_UPDATE.txt" | sed -n "s/^.*$DOWNLOAD_FIELD.*: \"\(.*\)\".*/\1/p"`
+DNL_LINK=`cat "$NZBOP_TEMPDIR/NZBGET_UPDATE.txt" | sed -n "s/^.*$DNL_FIELD.*: \"\(.*\)\".*/\1/p"`
+SIG_LINK=`cat "$NZBOP_TEMPDIR/NZBGET_UPDATE.txt" | sed -n "s/^.*$SIG_FIELD.*: \"\(.*\)\".*/\1/p"`
 rm -f "$NZBOP_TEMPDIR/NZBGET_UPDATE.txt"
 
-INSTALLER="nzbget-$VER-bin-linux.run"
-echo "Downloading $INSTALLER..."
-rm -f "$NZBOP_TEMPDIR/$INSTALLER"
-"$NZBOP_APPBIN" -B webget "$NZBOP_TEMPDIR/$INSTALLER" "$DOWNLOAD_LINK" 2>/dev/null
+SIGNATURE="nzbget-$VER.sig.txt"
+echo "Downloading verification signature..."
+rm -f "$NZBOP_TEMPDIR/$SIGNATURE"
+"$NZBOP_APPBIN" -n -B webget "$NZBOP_TEMPDIR/$SIGNATURE" "$SIG_LINK" 2>/dev/null
 if test "$?" != "0"; then
     echo "[ERROR] Download failed, please try again later"
     exit 1
 fi
+
+INSTALLER="nzbget-$VER-bin-linux.run"
+echo "Downloading $INSTALLER..."
+rm -f "$NZBOP_TEMPDIR/$INSTALLER"
+"$NZBOP_APPBIN" -n -B webget "$NZBOP_TEMPDIR/$INSTALLER" "$DNL_LINK" 2>/dev/null
+if test "$?" != "0"; then
+    echo "[ERROR] Download failed, please try again later"
+    exit 1
+fi
+
+echo "Verifying package authenticity..."
+"$NZBOP_APPBIN" -n -B verify "$NZBOP_APPDIR/pubkey.pem" "$NZBOP_TEMPDIR/$SIGNATURE" "$NZBOP_TEMPDIR/$INSTALLER" 2>/dev/null
+if test "$?" != "93"; then
+    echo "[ERROR] Package authenticity verification failed"
+    rm -f "$NZBOP_TEMPDIR/$INSTALLER"
+    rm -f "$NZBOP_TEMPDIR/$SIGNATURE"
+    exit 1
+fi
+rm -f "$NZBOP_TEMPDIR/$SIGNATURE"
 
 echo "Updating NZBGet..."
 echo "..."
